Visa banor uppladdade i systemet. Det går att tabort banor som inte används. Visa cykelvyn

+++ api/src/Domain/Model/Event/Service/EventService.php
@@ -2,22 +2,43 @@
 
 namespace App\Domain\Model\Event\Service;
 
+use App\common\Exceptions\BrevetException;
 use App\common\Service\ServiceAbstract;
+use App\Domain\Model\CheckPoint\Service\CheckpointsService;
 use App\Domain\Model\Event\Event;
 use App\Domain\Model\Event\Repository\EventRepository;
 use App\Domain\Model\Event\Rest\EventAssembly;
+use App\Domain\Model\Event\Rest\EventInformationAssembly;
+use App\Domain\Model\Event\Rest\EventInformationRepresentation;
 use App\Domain\Model\Event\Rest\EventRepresentation;
+use App\Domain\Model\Partisipant\Service\ParticipantService;
+use App\Domain\Model\Stats\Repository\StatisticsRepository;
+use App\Domain\Model\Stats\TrackStatistics;
+use App\Domain\Model\Track\Rest\TrackInformationAssembly;
+use App\Domain\Model\Track\Service\TrackService;
 use App\Domain\Permission\PermissionRepository;
 use Psr\Container\ContainerInterface;
 
 class EventService extends ServiceAbstract
 {
 
-    public function __construct(ContainerInterface $c, EventRepository $eventRepository, PermissionRepository $permissionRepository, EventAssembly $eventAssembly)
+    public function __construct(ContainerInterface $c, EventRepository $eventRepository,
+                                PermissionRepository $permissionRepository,
+                                EventAssembly $eventAssembly,
+                                EventInformationAssembly $eventInformationAssembly,
+                                TrackService $trackService,
+                                CheckpointsService $checkpointsService, ParticipantService $participantService,
+                                TrackInformationAssembly $trackInformationAssembly, StatisticsRepository $statisticsRepository)
     {
         $this->eventRepository = $eventRepository;
         $this->permissinrepository = $permissionRepository;
         $this->eventAssembly = $eventAssembly;
+        $this->eventInformationAssembly = $eventInformationAssembly;
+        $this->trackservice = $trackService;
+        $this->checkpointsService = $checkpointsService;
+        $this->participantService = $participantService;
+        $this->trackInformationAssembly = $trackInformationAssembly;
+        $this->statisticsRepository = $statisticsRepository;
     }
 
     public function allEvents(string $currentUserUid): array {
@@ -54,9 +75,17 @@ class EventService extends ServiceAbstract
         return $this->eventAssembly->toRepresentation($event,$permissions);
     }
 
-    public function deleteEvent(string $event_uid)
+    public function deleteEvent(string $event_uid, string $currentUserUid)
     {
-        $this->eventRepository->deleteEvent($event_uid);
+
+       $tracks = $this->trackservice->tracksForEvent($currentUserUid, $event_uid);
+
+       if($tracks == null || count($tracks) == 0){
+           $this->eventRepository->deleteEvent($event_uid);
+       } else {
+           throw new BrevetException("Det finns banor kopplade till eventet. Banorna måste tas bort från eventet", 5, null);
+       }
+
     }
 
     private function toEvent(EventRepresentation $eventRepresentation): Event {
@@ -78,5 +107,38 @@ class EventService extends ServiceAbstract
     {
         return $this->permissinrepository->getPermissionsTodata("EVENT",$user_uid);
 
+    }
+
+    public function eventInformation(string $eventUid, string $currentUserUid): array
+    {
+        $permissions = $this->getPermissions($currentUserUid);
+        $eventInfos = array();
+        $tracksarray = array();
+
+        if($eventUid != ""){
+            $event = $this->eventRepository->eventFor($eventUid);
+            $tracks = $this->trackservice->tracksForEvent($currentUserUid, $event->getEventUid());
+            foreach ($tracks as $track ){
+               // $participants = $this->participantService->participantsOnTrack($track->getTrackUid(), $currentUserUid);
+                $trackstats = $this->statisticsRepository->statsForTrack($track->getTrackUid());
+                array_push($tracksarray, $this->trackInformationAssembly->toRepresentation($this->trackservice->getTrackByTrackUid($track->getTrackUid(),$currentUserUid),$permissions, $currentUserUid,$trackstats == null ? null: $trackstats));
+            }
+                array_push($eventInfos, $this->eventInformationAssembly->toRepresentation($event, $tracksarray, $permissions, $currentUserUid));
+        } else {
+
+            $events = $this->eventRepository->allEvents();
+            foreach ($events as $event){
+                $tracksarray = [];
+                $tracks = $this->trackservice->tracksForEvent($currentUserUid, $event->getEventUid());
+                foreach ($tracks as $track ){
+                    //$participants = $this->participantService->participantsOnTrack($track->getTrackUid(), $currentUserUid);
+                    $trackstats = $this->statisticsRepository->statsForTrack($track->getTrackUid());
+                    array_push($tracksarray, $this->trackInformationAssembly->toRepresentation($this->trackservice->getTrackByTrackUid($track->getTrackUid(),$currentUserUid),$permissions, $currentUserUid,$trackstats == null ? null: $trackstats));
+                }
+                array_push($eventInfos, $this->eventInformationAssembly->toRepresentation($event, $tracksarray, $permissions, $currentUserUid));
+            }
+
+        }
+        return $eventInfos;
     }
 }
